feat: fixed the virusscan enabling problem. Moved the virusscan service to the document api configuration with ConditionalOnProperty
